refactor: 경험 단건 조회 위치 수정
